requirements.txt
asgiref
Django
sqlparse
pillow
psycopg2-binary
djangorestframework
rest_framework_simplejwt
django-filter
coverage
PyJWT
pytz
sqlparse
typing_extensions
tzdata
drf-yasg
django-cors-headers
requests
